+++ packages/core/utils/src/common/__tests__/define-config.spec.ts
@@ -1100,7 +1100,7 @@ describe("defineConfig", function () {
                 },
               ],
             },
-            "resolve": "@medusajs/caching",
+            "resolve": "@medusajs/medusa/caching",
           },
           "cart": {
             "resolve": "@medusajs/medusa/cart",
@@ -1114,6 +1114,9 @@ describe("defineConfig", function () {
           "event_bus": {
             "options": {
               "redisUrl": "redis://localhost:6379",
+              "workerOptions": {
+                "concurrency": 3,
+              },
             },
             "resolve": "@medusajs/medusa/event-bus-redis",
           },
@@ -1324,7 +1327,7 @@ describe("defineConfig", function () {
                 },
               ],
             },
-            "resolve": "@medusajs/caching",
+            "resolve": "@medusajs/medusa/caching",
           },
           "cart": {
             "resolve": "@medusajs/medusa/cart",
@@ -1338,6 +1341,9 @@ describe("defineConfig", function () {
           "event_bus": {
             "options": {
               "redisUrl": "redis://localhost:6379",
+              "workerOptions": {
+                "concurrency": 3,
+              },
             },
             "resolve": "@medusajs/medusa/event-bus-redis",
           },
@@ -1564,7 +1570,7 @@ describe("defineConfig", function () {
                 },
               ],
             },
-            "resolve": "@medusajs/caching",
+            "resolve": "@medusajs/medusa/caching",
           },
           "cart": {
             "resolve": "@medusajs/medusa/cart",
@@ -1578,6 +1584,9 @@ describe("defineConfig", function () {
           "event_bus": {
             "options": {
               "redisUrl": "redis://localhost:6379",
+              "workerOptions": {
+                "concurrency": 3,
+              },
             },
             "resolve": "@medusajs/medusa/event-bus-redis",
           },
@@ -2013,18 +2022,23 @@ describe("defineConfig", function () {
 
   it("should add cloud options to the project config and relevant modules if the environment variables are set", function () {
     const originalEnv = { ...process.env }
+    process.env.MEDUSA_BACKEND_URL = "test-backend-url"
     process.env.MEDUSA_CLOUD_ENVIRONMENT_HANDLE = "test-environment"
     process.env.MEDUSA_CLOUD_API_KEY = "test-api-key"
     process.env.MEDUSA_CLOUD_EMAILS_ENDPOINT = "test-emails-endpoint"
     process.env.MEDUSA_CLOUD_PAYMENTS_ENDPOINT = "test-payments-endpoint"
     process.env.MEDUSA_CLOUD_WEBHOOK_SECRET = "test-webhook-secret"
+    process.env.MEDUSA_CLOUD_OAUTH_AUTHORIZE_ENDPOINT =
+      "test-oauth-authorize-endpoint"
+    process.env.MEDUSA_CLOUD_OAUTH_TOKEN_ENDPOINT = "test-oauth-token-endpoint"
+    process.env.MEDUSA_CLOUD_OAUTH_DISABLED = "true"
     const config = defineConfig()
     process.env = { ...originalEnv }
 
     expect(config).toMatchInlineSnapshot(`
       {
         "admin": {
-          "backendUrl": "/",
+          "backendUrl": "test-backend-url",
           "path": "/app",
         },
         "featureFlags": {},
@@ -2035,6 +2049,15 @@ describe("defineConfig", function () {
           },
           "auth": {
             "options": {
+              "cloud": {
+                "api_key": "test-api-key",
+                "callback_url": "test-backend-url/app/login?auth_provider=cloud",
+                "disabled": true,
+                "environment_handle": "test-environment",
+                "oauth_authorize_endpoint": "test-oauth-authorize-endpoint",
+                "oauth_token_endpoint": "test-oauth-token-endpoint",
+                "sandbox_handle": undefined,
+              },
               "providers": [
                 {
                   "id": "emailpass",
@@ -2176,6 +2199,10 @@ describe("defineConfig", function () {
             "apiKey": "test-api-key",
             "emailsEndpoint": "test-emails-endpoint",
             "environmentHandle": "test-environment",
+            "oauthAuthorizeEndpoint": "test-oauth-authorize-endpoint",
+            "oauthCallbackUrl": undefined,
+            "oauthDisabled": true,
+            "oauthTokenEndpoint": "test-oauth-token-endpoint",
             "paymentsEndpoint": "test-payments-endpoint",
             "sandboxHandle": undefined,
             "webhookSecret": "test-webhook-secret",
@@ -2205,20 +2232,25 @@ describe("defineConfig", function () {
     `)
   })
 
-  it("should add cloud options to the project config and relevant modules if the environment varianbles is set for a sandbox", function () {
+  it("should add cloud options to the project config and relevant modules if the environment variable is set for a sandbox", function () {
     const originalEnv = { ...process.env }
+    process.env.MEDUSA_BACKEND_URL = "test-backend-url"
     process.env.MEDUSA_CLOUD_SANDBOX_HANDLE = "test-sandbox"
     process.env.MEDUSA_CLOUD_API_KEY = "test-api-key"
     process.env.MEDUSA_CLOUD_EMAILS_ENDPOINT = "test-emails-endpoint"
     process.env.MEDUSA_CLOUD_PAYMENTS_ENDPOINT = "test-payments-endpoint"
     process.env.MEDUSA_CLOUD_WEBHOOK_SECRET = "test-webhook-secret"
+    process.env.MEDUSA_CLOUD_OAUTH_AUTHORIZE_ENDPOINT =
+      "test-oauth-authorize-endpoint"
+    process.env.MEDUSA_CLOUD_OAUTH_TOKEN_ENDPOINT = "test-oauth-token-endpoint"
+    process.env.MEDUSA_CLOUD_OAUTH_DISABLED = "true"
     const config = defineConfig()
     process.env = { ...originalEnv }
 
     expect(config).toMatchInlineSnapshot(`
       {
         "admin": {
-          "backendUrl": "/",
+          "backendUrl": "test-backend-url",
           "path": "/app",
         },
         "featureFlags": {},
@@ -2229,6 +2261,15 @@ describe("defineConfig", function () {
           },
           "auth": {
             "options": {
+              "cloud": {
+                "api_key": "test-api-key",
+                "callback_url": "test-backend-url/app/login?auth_provider=cloud",
+                "disabled": true,
+                "environment_handle": undefined,
+                "oauth_authorize_endpoint": "test-oauth-authorize-endpoint",
+                "oauth_token_endpoint": "test-oauth-token-endpoint",
+                "sandbox_handle": "test-sandbox",
+              },
               "providers": [
                 {
                   "id": "emailpass",
@@ -2370,6 +2411,10 @@ describe("defineConfig", function () {
             "apiKey": "test-api-key",
             "emailsEndpoint": "test-emails-endpoint",
             "environmentHandle": undefined,
+            "oauthAuthorizeEndpoint": "test-oauth-authorize-endpoint",
+            "oauthCallbackUrl": undefined,
+            "oauthDisabled": true,
+            "oauthTokenEndpoint": "test-oauth-token-endpoint",
             "paymentsEndpoint": "test-payments-endpoint",
             "sandboxHandle": "test-sandbox",
             "webhookSecret": "test-webhook-secret",
@@ -2415,6 +2460,9 @@ describe("defineConfig", function () {
           webhookSecret: "overriden-webhook-secret",
           emailsEndpoint: "overriden-emails-endpoint",
           paymentsEndpoint: "overriden-payments-endpoint",
+          oauthAuthorizeEndpoint: "overriden-oauth-authorize-endpoint",
+          oauthTokenEndpoint: "overriden-oauth-token-endpoint",
+          oauthDisabled: true,
         },
       },
     })
@@ -2434,6 +2482,15 @@ describe("defineConfig", function () {
           },
           "auth": {
             "options": {
+              "cloud": {
+                "api_key": "overriden-api-key",
+                "callback_url": "//app/login?auth_provider=cloud",
+                "disabled": true,
+                "environment_handle": "overriden-environment",
+                "oauth_authorize_endpoint": "overriden-oauth-authorize-endpoint",
+                "oauth_token_endpoint": "overriden-oauth-token-endpoint",
+                "sandbox_handle": undefined,
+              },
               "providers": [
                 {
                   "id": "emailpass",
@@ -2575,6 +2632,10 @@ describe("defineConfig", function () {
             "apiKey": "overriden-api-key",
             "emailsEndpoint": "overriden-emails-endpoint",
             "environmentHandle": "overriden-environment",
+            "oauthAuthorizeEndpoint": "overriden-oauth-authorize-endpoint",
+            "oauthCallbackUrl": undefined,
+            "oauthDisabled": true,
+            "oauthTokenEndpoint": "overriden-oauth-token-endpoint",
             "paymentsEndpoint": "overriden-payments-endpoint",
             "sandboxHandle": undefined,
             "webhookSecret": "overriden-webhook-secret",