+++ www/packages/docs-ui/src/global-config.ts
@@ -4,7 +4,7 @@ import { DocsConfig } from "types"
 
 export const globalConfig: Pick<DocsConfig, "version"> = {
   "version": {
-    "number": "2.12.3",
-    "releaseUrl": "https://github.com/medusajs/medusa/releases/tag/v2.12.3"
+    "number": "2.12.5",
+    "releaseUrl": "https://github.com/medusajs/medusa/releases/tag/v2.12.5"
   }
 }